requirements.txt
gymnasium
numpy
torch
tensorboard
stable-baselines3
PyOpenGL
